requirements.txt
boto3 @ file:///home/conda/feedstock_root/build_artifacts/boto3_1747293660452/work
botocore @ file:///home/conda/feedstock_root/build_artifacts/botocore_1747285969285/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1687884292427/work
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560622100/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography-split_1746242003943/work
getmac==0.9.5
jmespath @ file:///home/conda/feedstock_root/build_artifacts/jmespath_1733229141657/work
mysql-connector-python==9.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
ply @ file:///home/conda/feedstock_root/build_artifacts/ply_1733239724146/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
PyQt5-sip==12.9.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
s3transfer @ file:///home/conda/feedstock_root/build_artifacts/s3transfer_1745398507755/work
sip @ file:///Users/runner/miniforge3/conda-bld/sip_1681161119931/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1733256695513/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1718728347128/work
